Speed improvement
When trying to match the league, all leagues are collected from TSDB. In
that data we can access the 'standard name' of the league. Instead of
making an API call for each league at this step attempt to see if they
are a "perfect" match.

If that fails then make an API call for each league and try to match
against alternate names.

If a match is found break the loop. I don't think there is any reason to
keep searching?